==> notebooks/Step 1. Filter Charges.ipynb <==
# This is the first step in the analysis: Filter cases to the ones in which we are interested

import pandas as pd
import numpy as np
import datetime
import csv

crim_charge = pd.read_excel("../data/inputs/ChargeReport12_04_2017.xlsx")
crim_charge.head()

## Sub-step 1: Filter to cases post-2000

date_filtered_case_nums = set(
    [
        case[1]['Case No']
        for case in crim_charge.iterrows()
        if case[1]['Offense Date'] > pd.Timestamp('2000-01-01 00:00:00')
    ]
)
len(date_filtered_case_nums)

## Sub-step 2: Filter to cases with some elder abuse charge

def is_elder_case(charge):
    elder_charge_keywords = ('elder', '60', '65')
    if any(kw in charge[1]['Offense Description'].lower() for kw in elder_charge_keywords):
        return True
    return False

elder_case_nums = set(
    [charge[1]['Case No'] for charge in crim_charge.iterrows() if is_elder_case(charge)]
)

## Sub-step 3: Combine filters

filtered_charges = [
    charge[1] for charge in crim_charge.iterrows()
    if charge[1]['Case No'] in elder_case_nums
    and charge[1]['Case No'] in date_filtered_case_nums
]

to_write = pd.DataFrame(filtered_charges)

## Sub-step 4: Write data to `steps` folder

writer = pd.ExcelWriter('../data/steps/Step1Output_charges.xlsx')
to_write.to_excel(writer,'Sheet1')
writer.save()
