==> notebooks/Step 2. Merge Sentence Data.ipynb <==
# This step will merge the sentence data we received in Dec. 2017 with new sentence data from May 2018

import pandas as pd
import numpy as np
import datetime
import csv

old_sentences = pd.read_excel("../data/inputs/Sentences_12_04_2017.xlsx")
old_sentences.head()

new_sentences = pd.read_excel("../data/inputs/Sentences_New_5_6_2018.xlsx")
new_sentences.head()

## Sub-step 1: Create mappings from old sentencing charge keys (Case Number + Charge Number) to old sentence record and vice versa

old_sentence_mapping = {}

#### There are some collisions in the old sentencing data. Want to use the most recent record every time

def get_max_sentence_record(key, row):
    if key in old_sentence_mapping:
        return max(
            row[1],
            old_sentence_mapping[key],
            key=lambda x:x['SentenceDate']
        )
    return row[1]

for row in old_sentences.iterrows():
    key = (row[1]['CaseNumber'], row[1]['ChargeNumber'])        
    old_sentence_mapping[key] = get_max_sentence_record(key, row)

#### Need to make 2 mappings, one for the `Case Number` column in the new data and one for the `Associated Case` column

new_dc_sentence_mapping = {}
new_sup_sentence_mapping = {}

all_keys = []
for row in new_sentences.iterrows():
    all_keys.append(
        (
            row[1]['Case Number'], 
            row[1]['Associated Case'],
            row[1]['Charge Number']
        )
    )
    dc_key = (row[1]['Case Number'], row[1]['Charge Number'])
    sup_key = (row[1]['Associated Case'], row[1]['Charge Number'])
    new_dc_sentence_mapping[dc_key] = row[1]
    new_sup_sentence_mapping[sup_key] = row[1]

## Sub-step 2: Ensure new sentencing information adds no new charges, only updates existing charges

#### We want to confirm that every charge ID in the new sentencing data is present in the old sentencing data to simplify analysis

old_only = []
new_only = []
both = []

for key in all_keys:
    dc_key = (key[0], key[2])
    sup_key = (key[1], key[2])
    if not (dc_key in old_sentence_mapping or sup_key in old_sentence_mapping):
        new_only.append(row[1])
print 'There are {} charges that are only in the new data.'.format(len(new_only))

#### So our assertion is incorrect; there are some sentences in the new data that didn't exist in the old. Now, let's make sure they a) aren't Null and b) are elder abuse

for k in new_only:
    p = ('elder', '65', '60')
    if pd.isnull(k['Sentence Conditions']):
        continue
    if any(o in k['Sentence Conditions'] for o in p):
        print k

#### They are all Null/not elder, so we can ignore them. Therefore, the old sentencing data contains all relevant keys.

## Sub-step 3: Create merged sentencing data, using new sentencing data if available, and old otherwise

merged_sentences = []

def get_sentence_info(new, old):
    merged_sentence = {
        'charge_id': old['ChargeID']
    }
    if new is not None:
        merged_sentence.update({
            'disposition': new['Disposition'],
            'disposition_date': new['Disposition Date'],
            'sentence': new['Sentence'],
            'sentence_date': new['Sentence Date'],
            'sentence_terms': new['Sentence Terms'],
            'sentence_conditions': new['Sentence Conditions'],
            'judge': new['Judge']
        })
    else:
        merged_sentence.update({
            'disposition': old['Disposition'],
            'disposition_date': old['DispositionDate'],
            'sentence': old['Sentence'],
            'sentence_date': old['SentenceDate'],
            'sentence_terms': old['SentenceTerms'],
            'sentence_conditions': old['SentenceConditions'],
            'judge': old['Judge']
        })    
    return merged_sentence

for key in old_sentence_mapping:
    old = old_sentence_mapping[key]
    new = None
    if key in new_dc_sentence_mapping:
        new = new_dc_sentence_mapping[key]
    elif key in new_sup_sentence_mapping:
        new = new_sup_sentence_mapping[key]
    merged_sentences.append(get_sentence_info(new, old))

len(merged_sentences)

to_write = pd.DataFrame(merged_sentences)

## Sub-step 4: Write data to `steps` folder

writer = pd.ExcelWriter('../data/steps/Step2Output_sentences.xlsx')
to_write.to_excel(writer,'Sheet1')
writer.save()
