==> notebooks/Step 3. Add Sentences to Filtered Charges.ipynb <==
# Combine our filtered charges with our sentencing information

import pandas as pd
import numpy as np
import datetime
import csv

crim_sentence = pd.read_excel('../data/steps/Step2Output_sentences.xlsx')
crim_sentence.head()

filtered_charges = pd.read_excel('../data/steps/Step1Output_charges.xlsx')
filtered_charges.head()

## Sub-step 1: Ensure all filtered charges have a matching sentence

#### First, let's create a mapping from Charge ID to sentence

charge_to_sentence_map = {}

for row in crim_sentence.iterrows():
    charge_to_sentence_map[row[1]['charge_id']] = row[1]

#### Now, let's make sure every filtered charge has a matching sentence

total = 0
elder_unmatched = 0
matched = []
for row in filtered_charges.iterrows():
    if row[1]['Charge ID'] in charge_to_sentence_map:
        matched.append(row[1])
    else:
        elder_kws = ('elder','60','65')
        if any(kw in row[1]['Offense Description'] for kw in elder_kws):
            elder_unmatched += 0
    total += 1

print '{}/{} filtered charges could be mapped by Charge ID'.format(len(matched),total)
print '{}/{} filtered charges could NOT be mapped by Charge ID, but only {} of those were elder charges'.format(
    total-len(matched),total,elder_unmatched
)

#### Okay, so they're not all matchable by Charge ID, but only 8 aren't and they aren't even elder abuse charges, so good to go matching by charge ID

## Sub-step 2: Create full charge + sentencing record for filtered cases

def snake_case(string):
    return string.lower().replace(' ', '_')

def create_merged_blob(charge, sentence):
    blob = {}
    for key in charge.keys():
        blob[snake_case(key)] = charge[key]
    for key in sentence.keys():
        blob[snake_case(key)] = sentence[key]
    return blob

blobs = []
for charge in matched:
    sentence = charge_to_sentence_map[charge['Charge ID']]
    full_blob = create_merged_blob(charge, sentence)
    blobs.append(full_blob)

len(blobs)

to_write = pd.DataFrame(blobs)

## Sub-step 3: Write data to `steps` folder

writer = pd.ExcelWriter('../data/steps/Step3Output_charges_sentences.xlsx')
to_write.to_excel(writer,'Sheet1')
writer.save()
