==> notebooks/Step 4. Add Defendant Data.ipynb <==
# Add the defendant information to each charge

import pandas as pd
import numpy as np
import datetime
import csv

crim_defendant = pd.read_excel('../data/inputs/Defendant_12_04_2017.xlsx')
crim_defendant.head()

charges_with_sentences = pd.read_excel('../data/steps/Step3Output_charges_sentences.xlsx')
charges_with_sentences.head()

## Sub-step 1: Create a mapping from case number to defendant data

defendant_mapping = {}
for row in crim_defendant.iterrows():
    case_number = row[1]['Case Number']
    defendant_mapping[case_number] = row[1]

### Sub-step 2: Merge the charge data and defendant data

def snake_case(string):
    return string.lower().replace(' ', '_')

def create_merged_blob(charge, defendant):
    blob = {}
    for key in charge.keys():
        blob[snake_case(key)] = charge[key]
    for key in defendant.keys():
        if key == 'Case Number':
            continue
        blob[snake_case(key)] = defendant[key]
    return blob

merged_data = []
for row in charges_with_sentences.iterrows():
    case_no = row[1]['case_no']
    merged_blob = create_merged_blob(
        row[1],
        defendant_mapping[case_no]
    )
    merged_data.append(merged_blob)

to_write = pd.DataFrame(merged_data)
len(to_write)

## Sub-step 3: Write data to `steps` folder

writer = pd.ExcelWriter('../data/steps/Step4Output_merged.xlsx')
to_write.to_excel(writer,'Sheet1')
writer.save()
