==> notebooks/Step 5. Add District & Superior Case Number Links.ipynb <==
# Combine district court cases with their superior court counterparts

import pandas as pd
import numpy as np
import datetime
import csv

charge_data = pd.read_excel('../data/steps/Step4Output_merged.xlsx')
charge_data.head()

## Sub-step 1: Combine DC/Superior linkage files to create case mappings

dc_to_sup = {}
sup_to_dc = {}
dc_to_dc = {}
sup_to_sup = {}

def add_to_set_map(case_map, key, value):
    if key in case_map:
        case_map[key].add(value)
    else:
        case_map[key] = set()
        case_map[key].add(value)

with open('../data/inputs/dc_links.csv') as dc_links, open('../data/inputs/sup_links.csv') as sup_links:
    dc_reader = csv.reader(dc_links)
    sup_reader = csv.reader(sup_links)
    next(dc_reader)
    next(sup_reader)
    for line in dc_reader:
        if line[1] == line[3]:
            add_to_set_map(dc_to_dc, line[0], line[2])
            add_to_set_map(dc_to_dc, line[0], line[0]) # Always map an item to itself for dc to dc
        elif line[2] != 'NULL':
            add_to_set_map(dc_to_sup, line[0], line[2])
    for line in sup_reader:
        if line[1] == line[3]:
            add_to_set_map(sup_to_sup, line[0], line[2])
            add_to_set_map(sup_to_sup, line[0], line[0]) # Always map an item to itself for sup to sup
        elif line[2] != 'NULL':
            add_to_set_map(sup_to_dc, line[0], line[2])

## Sub-step 2: Get a list of all of our case numbers (and make a mapping of case number -> list of charges while we're at it)

def add_to_list_map(case_map, key, value):
    if key in case_map:
        case_map[key].append(value)
    else:
        case_map[key] = []
        case_map[key].append(value)

charge_map = {}
for row in charge_data.iterrows():
    case_number = row[1]['case_no']
    add_to_list_map(charge_map, case_number, row[1])

## Sub-step 3: Create unique versions of `dc_to_dc` and `sup_to_sup` that are specific to our cases, and can be merged easily

unique_sup_to_sup = []
unique_dc_to_dc = []

sup_handled = set()
dc_handled = set()
for key in charge_map:
    if key in sup_to_sup and not key in sup_handled:
        value = [v for v in sup_to_sup[key] if v in charge_map]
        if len(value) <= 1:
            continue
        unique_sup_to_sup.append(value)
        for v in value:
            sup_handled.add(v)
    if key in dc_to_dc and not key in dc_handled:
        value = [v for v in dc_to_dc[key] if v in charge_map]
        if len(value) <= 1:
            continue
        unique_dc_to_dc.append(value)
        for v in value:
            dc_handled.add(v)
print(len(unique_sup_to_sup))

#### These now contain lists of groups of case numbers that are related. Some are actually dc->sup mappings, so we'll handle those

def get_dc_to_sup(group):
    # Returns dc, sup if True
    if len(group) != 2:
        return None, None
    item1 = group[0]
    item2 = group[1]
    if item1[0].isalpha() and not item2[0].isalpha():
        return item2, item1
    if not item1[0].isalpha() and item2[0].isalpha():
        return item1, item2
    return None, None

for group in unique_sup_to_sup:
    dc, sup = get_dc_to_sup(group)
    if dc and sup:
        unique_sup_to_sup.remove(group)
        add_to_set_map(dc_to_sup, dc, sup)
print(len(unique_sup_to_sup))

#### Now we've just got weird case number mappings. So for these, pick the one with the higher charge date, and remove the other from our data mapping -- they should have the same data anyway

def get_max(group, field, full=False): # Update this to check vals
    max_field = None
    max_case_num = None
    
    sorted(
        group,
        key=lambda x:max(charge_map[x],key=lambda x:x[field])[field],
        reverse=True
    )
    if full:
        return group
    return group[0]

for group in unique_sup_to_sup:
    max_case_num = get_max(group, 'charge_filing_date')
    for ele in group:
        if ele != max_case_num:
            del charge_map[ele]
print(len(charge_map))

#### Do the same for dc_to_dc

for group in unique_dc_to_dc:
    max_case_num = get_max(group, 'charge_filing_date')
    for ele in group:
        if ele != max_case_num:
            del charge_map[ele]
print(len(charge_map))

## Sub-step 4: use `dc_to_sup` to combine dc and sup data

def check_linked_case_num(linked_case_number):
    linked_case_num = next(iter(linked_case_number))
    if linked_case_num in charge_map:
        return linked_case_num
    return None

def filter_multiple_linked_cases(current_case_number, linked_case_nums):
    if current_case_number[0].isalpha():        
        return get_max(
            [
                case_number for case_number in linked_case_nums
                if case_number in charge_map and not case_number[0].isalpha()
            ], 'charge_filing_date', full=True
        )
    else:
        return get_max(
            [
                case_number for case_number in linked_case_nums
                if case_number in charge_map and case_number[0].isalpha()
            ], 'charge_filing_date', full=True
        )

approved_cases = {}
ignore = set()
deleted = 0
ignoring = 0
case_nums = set()
for row in charge_data.iterrows():
    case_number = row[1]['case_no']
    if case_number not in charge_map:
        deleted +=1
        continue
    if case_number in ignore:
        ignoring += 1
        continue
    case_nums.add(case_number)
    
    linked_case_num = None
    if case_number in dc_to_sup or case_number in sup_to_dc:
        if case_number[0].isalpha():
            linked_case_numbers = sup_to_dc[case_number]
        else:
            linked_case_numbers = dc_to_sup[case_number]

        if len(linked_case_numbers) == 1:
            linked_case_num = check_linked_case_num(linked_case_numbers)
        else:
            ordered_linked_case_nums = filter_multiple_linked_cases(case_number, linked_case_numbers)
            if len(ordered_linked_case_nums):
                linked_case_num = ordered_linked_case_nums[0]
            else:
                linked_case_num = None
            for k in ordered_linked_case_nums[1:]:
                ignore.add(k)

    if linked_case_num is None:
        approved_cases[case_number] = None
    else:
        ordered = get_max([case_number, linked_case_num], 'charge_filing_date', full=True)
        ignore.add(ordered[1])
        if ordered[0] == case_number:
            approved_cases[case_number] = ordered[1]
print(deleted, ignoring)

to_write = []
for row in charge_data.iterrows():
    charge_record = row[1]
    case_number = charge_record['case_no']

    if case_number in approved_cases:
        if case_number[0].isalpha():
            charge_record['sup_case'] = case_number
            charge_record['dc_case'] = approved_cases[case_number]
        else:
            charge_record['sup_case'] = approved_cases[case_number]
            charge_record['dc_case'] = case_number
        del row[1]['case_no']
        to_write.append(charge_record)

print(len(to_write))

## Sub-step 5: Write data to `steps` folder

writer = pd.ExcelWriter('../data/steps/Step5Output_joined.xlsx')
pd.DataFrame(to_write).to_excel(writer,'Sheet1')
writer.save()
